#9476 DI injection via method signature. refactoring unit test.
