+++ tests/framework/web/stubs/Bar.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yiiunit\framework\web\stubs;
+
+use yii\base\Object;
+
+/**
+ * @author Misbahul D Munir <misbahuldmunir@gmail.com>
+ * @since 2.0
+ */
+class Bar extends Object
+{
+    public $foo;
+}
